Blockades
- Added BlockadeScript & HealthScript for Blockades
> BlockadeGroups prefabs now use a parent Hitbox collider instead of each individual "blockade" object having its own colliders. This is because every object in a single blockade prefab shares the same health pool, therefore having a parent hitbox eliminates hitting multiple objects at once causing multiple damages on hit
- Implemented randomized blockade spawning in LevelManager

TODO:
- Item drop on death

+++ Assets/Scripts/_Datas/PlayerData (GameData)/Ammo.cs
@@ -15,7 +15,7 @@ public class Ammo
     public float amountPerStack;
     public float price;
 
-    public Ammo(AmmoType ammoType, int amount, float maxAmount, float amountPerStack, float price)
+    public Ammo(AmmoType ammoType, float amount, float maxAmount, float amountPerStack, float price)
     {
         this.ammoType = ammoType;
         this.amount = Math.Clamp(amount, 0, maxAmount);
